Fixed delete key and more button issues
With multi characters (where letters combine to form a new letter) you
can again delete properly, deleting each part of the letter rather then
simply deleting the letter as a whole.

Also fixed the keyboard shortcut so many people seem to use where you
can drag from the "123" button to a number and it will return to the
letter keyboard.